file handling complete

+++ functions.py
@@ -1,12 +1,10 @@
-
+import os
 def readToDO(filepath='../app1/files/toDos.txt'):
-	try:
-		with open(filepath, 'r') as file:
-			toDos = file.readlines()
-			return toDos
-	except FileNotFoundError:
+	if not os.path.exists('../app1/files/toDos.txt'):
 		file = open('../app1/files/toDos.txt', 'w')
-
+	with open(filepath, 'r') as file:
+		toDos = file.readlines()
+		return toDos
 def writeToDo(toDos,filepath = '../app1/files/toDos.txt'):
 	with open(filepath, 'w') as file:
 		file.writelines(toDos)
